+++ pipeline/stepBuilder.go
@@ -17,7 +17,6 @@ package pipeline
 
 import (
 	"fmt"
-	"net/url"
 	"path/filepath"
 	"strings"
 
@@ -26,16 +25,15 @@ import (
 	"github.com/rs/zerolog/log"
 
 	backend "github.com/woodpecker-ci/woodpecker/pipeline/backend/types"
+	"github.com/woodpecker-ci/woodpecker/pipeline/frontend"
 	"github.com/woodpecker-ci/woodpecker/pipeline/frontend/metadata"
 	"github.com/woodpecker-ci/woodpecker/pipeline/frontend/yaml"
 	"github.com/woodpecker-ci/woodpecker/pipeline/frontend/yaml/compiler"
 	"github.com/woodpecker-ci/woodpecker/pipeline/frontend/yaml/linter"
 	"github.com/woodpecker-ci/woodpecker/pipeline/frontend/yaml/matrix"
 	"github.com/woodpecker-ci/woodpecker/server"
-	"github.com/woodpecker-ci/woodpecker/server/forge"
 	forge_types "github.com/woodpecker-ci/woodpecker/server/forge/types"
 	"github.com/woodpecker-ci/woodpecker/server/model"
-	"github.com/woodpecker-ci/woodpecker/version"
 )
 
 // StepBuilder Takes the hook data and the yaml and returns in internal data model
@@ -49,6 +47,7 @@ type StepBuilder struct {
 	Link  string
 	Yamls []*forge_types.FileMeta
 	Envs  map[string]string
+	Forge metadata.ServerForge
 }
 
 type Item struct {
@@ -88,8 +87,8 @@ func (b *StepBuilder) Build() ([]*Item, error) {
 			}
 
 			// TODO: move metadata into own package and inject as argument into Build()
-			metadata := MetadataFromStruct(server.Config.Services.Forge, b.Repo, b.Curr, b.Last, workflow, b.Link)
-			environ := b.environmentVariables(metadata, axis)
+			workflowMetadata := frontend.MetadataFromStruct(b.Forge, b.Repo, b.Curr, b.Last, workflow, b.Link)
+			environ := b.environmentVariables(workflowMetadata, axis)
 
 			// add global environment variables for substituting
 			for k, v := range b.Envs {
@@ -120,7 +119,7 @@ func (b *StepBuilder) Build() ([]*Item, error) {
 			}
 
 			// checking if filtered.
-			if match, err := parsed.When.Match(metadata, true); !match && err == nil {
+			if match, err := parsed.When.Match(workflowMetadata, true); !match && err == nil {
 				log.Debug().Str("pipeline", workflow.Name).Msg(
 					"Marked as skipped, dose not match metadata",
 				)
@@ -132,7 +131,7 @@ func (b *StepBuilder) Build() ([]*Item, error) {
 				return nil, err
 			}
 
-			ir, err := b.toInternalRepresentation(parsed, environ, metadata, workflow.ID)
+			ir, err := b.toInternalRepresentation(parsed, environ, workflowMetadata, workflow.ID)
 			if err != nil {
 				return nil, err
 			}
@@ -322,102 +321,6 @@ func SetPipelineStepsOnPipeline(pipeline *model.Pipeline, pipelineItems []*Item)
 	}
 
 	return pipeline
-}
-
-// MetadataFromStruct return the metadata from a pipeline will run with.
-func MetadataFromStruct(forge forge.Forge, repo *model.Repo, pipeline, last *model.Pipeline, workflow *model.Step, link string) metadata.Metadata {
-	host := link
-	uri, err := url.Parse(link)
-	if err == nil {
-		host = uri.Host
-	}
-
-	fForge := metadata.Forge{}
-	if forge != nil {
-		fForge = metadata.Forge{
-			Type: forge.Name(),
-			URL:  forge.URL(),
-		}
-	}
-
-	fRepo := metadata.Repo{}
-	if repo != nil {
-		fRepo = metadata.Repo{
-			Name:     repo.FullName,
-			Link:     repo.Link,
-			CloneURL: repo.Clone,
-			Private:  repo.IsSCMPrivate,
-			Branch:   repo.Branch,
-		}
-	}
-
-	fWorkflow := metadata.Workflow{}
-	if workflow != nil {
-		fWorkflow = metadata.Workflow{
-			Name:   workflow.Name,
-			Number: workflow.PID,
-			Matrix: workflow.Environ,
-		}
-	}
-
-	return metadata.Metadata{
-		Repo:     fRepo,
-		Curr:     metadataPipelineFromModelPipeline(pipeline, true),
-		Prev:     metadataPipelineFromModelPipeline(last, false),
-		Workflow: fWorkflow,
-		Step:     metadata.Step{},
-		Sys: metadata.System{
-			Name:     "woodpecker",
-			Link:     link,
-			Host:     host,
-			Platform: "", // will be set by pipeline platform option or by agent
-			Version:  version.Version,
-		},
-		Forge: fForge,
-	}
-}
-
-func metadataPipelineFromModelPipeline(pipeline *model.Pipeline, includeParent bool) metadata.Pipeline {
-	if pipeline == nil {
-		return metadata.Pipeline{}
-	}
-
-	cron := ""
-	if pipeline.Event == model.EventCron {
-		cron = pipeline.Sender
-	}
-
-	parent := int64(0)
-	if includeParent {
-		parent = pipeline.Parent
-	}
-
-	return metadata.Pipeline{
-		Number:   pipeline.Number,
-		Parent:   parent,
-		Created:  pipeline.Created,
-		Started:  pipeline.Started,
-		Finished: pipeline.Finished,
-		Status:   string(pipeline.Status),
-		Event:    string(pipeline.Event),
-		Link:     pipeline.Link,
-		Target:   pipeline.Deploy,
-		Commit: metadata.Commit{
-			Sha:     pipeline.Commit,
-			Ref:     pipeline.Ref,
-			Refspec: pipeline.Refspec,
-			Branch:  pipeline.Branch,
-			Message: pipeline.Message,
-			Author: metadata.Author{
-				Name:   pipeline.Author,
-				Email:  pipeline.Email,
-				Avatar: pipeline.Avatar,
-			},
-			ChangedFiles:      pipeline.ChangedFiles,
-			PullRequestLabels: pipeline.PullRequestLabels,
-		},
-		Cron: cron,
-	}
 }
 
 func SanitizePath(path string) string {
